Smart colony id command arguments (#11125)
All commands that take a colony id now accept additional values in its place:
The colony id, as before (but now it suggests colony ids that you've been close to recently, i.e. are still in the client-side colony manager; you can still type other ids)
The name or id of a player, which resolves to their owned colony (and it suggests known player names, but not ids)
@here, which means the colony where you're running the command
@mine, which means your own colony
IColonyManager.getColonies() now works on client side too.