Some necessary de-fuckery of the window logic (#11203)
General cleanup of some of the module window code (note: no actual modifications were made)
Firstly, all of the windows now are required to pass a proper ResourceLocation as the window identifier, no more strings!
All of the page identifiers were ripped out of constants, and all moved into the respective windows directly
Constructors of windows have all been made that the IBuildingModuleView or whatever implementation must be passed into the window as the only argument. Building views are obsolete as these can be obtained via the module view themselves anyway
Unhinged the root building window logic, there used to be 2 classes, the AbstractWindowModuleBuilding and the AbstractModuleWindow. The setup behind this was a little confusing. The AbstractWindowModuleBuilding was handling the "main" page of buildings, but was needing a module, which had to just be passed as null. I shuffled the logic around a bit more easily separated. There are now 3 classes: AbstractBuildingWindow which stores the building view, AbstractModuleWindow, which extends from the building window, and stores the module view, and AbstractBuildingMainWindow which defines all the "main" page code
All of the doubly storing of the building/module views has been removed, and are now using the protected fields from the superclasses
Made support for potentially rendering tab icons on either side of the page, via new methods in the superclass